refactor: split Config types into config.rs, make mod.rs a pure aggregator
Move the Config struct, ServerState, TileSourceWarning, OnInvalid,
parse_base_path, and init_aws_lc_tls from main/mod.rs into a sibling
main/config.rs along with the non-parse tests. mod.rs is now just five
lines wiring mod config / lifecycle / parse and re-exporting config::*
and parse::*.

No public-API change: crate::config::file::Config and friends remain at
the same path via the re-exports.
